# Optimizer: one LR op per (device, optimizer)