--- a/caffe2/python/optimizer.py
+++ b/caffe2/python/optimizer.py
@@ -44,10 +44,23 @@ class Optimizer(object):
     def _run(self, net, param_init_net, param_info):
         raise Exception("Not Impelemented")
 
-    @staticmethod
-    def build_lr(net, param_init_net, base_learning_rate,
-                 learning_rate_blob="lr", policy="fixed",
+    def get_lr_blob_name(self):
+        """Returns an LR blob name.
+        The name will be unique to the current device and optimizer type.
+        """
+        classname = self.__class__.__name__
+        s = scope.CurrentDeviceScope()
+        if s.device_type == caffe2_pb2.CUDA:
+            return '%s_lr_gpu%d' % (classname, s.cuda_gpu_id)
+        else:
+            return '%s_lr_cpu' % classname
+
+
+    def build_lr(self, net, param_init_net, base_learning_rate,
+                 learning_rate_blob=None, policy="fixed",
                  iter_val=0, **kwargs):
+        if learning_rate_blob is None:
+            learning_rate_blob = self.get_lr_blob_name()
         if not param_init_net.BlobIsDefined(_OPTIMIZER_ITERATION_NAME):
             # Add training operators.
             with core.DeviceScope(core.DeviceOption(caffe2_pb2.CPU)):
@@ -60,15 +73,18 @@ class Optimizer(object):
         else:
             iteration = param_init_net.GetBlobRef(_OPTIMIZER_ITERATION_NAME)
 
-        # There is one interesting thing here: since we are minimizing, we are
-        # doing "descent" so the learning rate is set to be negative.
-        lr = net.LearningRate(
-            [iteration],
-            learning_rate_blob,
-            base_lr=-base_learning_rate,
-            policy=policy,
-            **kwargs
-        )
+        if not net.BlobIsDefined(learning_rate_blob):
+            # There is one interesting thing here: since we are minimizing, we are
+            # doing "descent" so the learning rate is set to be negative.
+            lr = net.LearningRate(
+                [iteration],
+                learning_rate_blob,
+                base_lr=-base_learning_rate,
+                policy=policy,
+                **kwargs
+            )
+        else:
+            lr = net.GetBlobRef(learning_rate_blob)
         return lr, iteration
 
     @staticmethod
@@ -138,7 +154,6 @@ class SgdOptimizer(Optimizer):
         lr, _ = self.build_lr(
             net, param_init_net,
             base_learning_rate=self.base_learning_rate * lr_sign,
-            learning_rate_blob=str(param) + "_lr",
             policy=self.policy,
             **(self.init_kwargs)
         )
@@ -219,7 +234,6 @@ class MultiPrecisionSgdOptimizer(SgdOptimizer):
         lr, _ = self.build_lr(
             net, param_init_net,
             base_learning_rate=-self.base_learning_rate,
-            learning_rate_blob=param + "_lr",
             policy=self.policy,
             **(self.init_kwargs)
         )
@@ -295,7 +309,6 @@ class AdagradOptimizer(Optimizer):
         lr, _ = self.build_lr(
             net, param_init_net,
             base_learning_rate=self.alpha,
-            learning_rate_blob=str(param) + "_lr",
             policy=self.policy,
             **(self.init_kwargs)
         )
@@ -404,7 +417,6 @@ class AdamOptimizer(Optimizer):
         lr, iteration = self.build_lr(
             net, param_init_net,
             base_learning_rate=self.alpha,
-            learning_rate_blob=str(param) + "_lr",
             policy=self.policy,
             **(self.init_kwargs)
         )