@lukeyeager 2017-07-05
> Is there any particular reason you decided to have a separate LearningRate op for every parameter in 1317e34?

Ping.